requirements.txt
einops==0.6.1
matplotlib==3.7.1
numpy==1.25.0
pandas==2.0.3
patool==1.12
pytorch_lightning==2.0.4
PyYAML==6.0.1
scikit_learn==1.2.2
sktime==0.20.0
timm==0.9.2
torch==2.0.0
torchinfo==1.7.2
torchmetrics==0.11.4
tqdm==4.65.0
